Stream the unknown arenas the validators only scan
Three validation checks materialized every reserved `unknowns` record in the
document to read one field off each. Retained populations reach hundreds of
thousands of records, so `all_native_unknowns` allocated an id string and a link
vector per record and held the whole typed population while the check built its
own — a second population — from it. None of the three wants the records: the
carrier-reachability check wants link targets, the native-link check wants each
record once, and the referential-integrity id set wants ids.

`CadIr::all_native_unknowns_iter` is the streaming core, and
`all_native_unknowns` now collects from it, mirroring `native_unknowns_iter`.
`carriers_parameterization` folds the links out of the stream and keeps only
those, `annotations_native` consumes each record and drops it, and `topology`
collects ids straight into its presence set. The two fallible-in-bulk sites keep
their all-or-nothing reading: a `Result` collect and a `try_fold` still yield
nothing when an arena cannot be read, as `unwrap_or_default` on the vector did.
`annotations_native` instead skips only the records it cannot read, since it
reports per record rather than building one value.

+++ crates/cadmpeg-codec-f3d/src/nurbs/blend.rs
@@ -959,10 +959,10 @@ fn decode_vertex_blend_boundary(
     int_width: usize,
 ) -> Option<EmbeddedVertexBlendBoundary> {
     let kind = take_native_string(bytes, position)?;
-    let boundary_type = i64::from(take_bool(bytes, position)?);
+    let boundary_type = take_bool(bytes, position)?;
     let magic = take_native_vec3(bytes, position, 0x13)?;
-    let u_smoothing = i64::from(take_bool(bytes, position)?);
-    let v_smoothing = i64::from(take_bool(bytes, position)?);
+    let u_smoothing = take_bool(bytes, position)?;
+    let v_smoothing = take_bool(bytes, position)?;
     let fullness = take_f64(bytes, position)?;
     let geometry = match kind.as_str() {
         "circle" => {
@@ -985,7 +985,7 @@ fn decode_vertex_blend_boundary(
                 ));
             }
             let parameters = [take_f64(bytes, position)?, take_f64(bytes, position)?];
-            let sense = i64::from(take_bool(bytes, position)?);
+            let sense = take_bool(bytes, position)?;
             EmbeddedVertexBlendBoundaryGeometry::Circle {
                 curve: CurveGeometry::Nurbs(curve.curve),
                 curve_endpoints: [None; 2],
@@ -1014,7 +1014,7 @@ fn decode_vertex_blend_boundary(
         "pcurve" => {
             let surface = decode_embedded_surface(bytes, position, int_width)?;
             let pcurve = decode_nullable_embedded_pcurve(bytes, position, int_width)?;
-            let sense = i64::from(take_bool(bytes, position)?);
+            let sense = take_bool(bytes, position)?;
             let fit_tolerance = take_f64(bytes, position)?;
             EmbeddedVertexBlendBoundaryGeometry::Pcurve {
                 surface,
@@ -1062,10 +1062,10 @@ fn decode_revision_vertex_blend_boundary(
 ) -> Option<EmbeddedVertexBlendBoundary> {
     let (active_bytes, tables) = resolver?;
     let kind = take_native_ident(bytes, position)?;
-    let boundary_type = i64::from(take_bool(bytes, position)?);
+    let boundary_type = take_bool(bytes, position)?;
     let magic = take_native_vec3(bytes, position, 0x14)?;
-    let u_smoothing = i64::from(take_bool(bytes, position)?);
-    let v_smoothing = i64::from(take_bool(bytes, position)?);
+    let u_smoothing = take_bool(bytes, position)?;
+    let v_smoothing = take_bool(bytes, position)?;
     let fullness = take_f64(bytes, position)?;
     let geometry = match kind.as_str() {
         "circle" => {
@@ -1097,7 +1097,7 @@ fn decode_revision_vertex_blend_boundary(
                 ));
             }
             let parameters = [take_f64(bytes, position)?, take_f64(bytes, position)?];
-            let sense = i64::from(take_bool(bytes, position)?);
+            let sense = take_bool(bytes, position)?;
             EmbeddedVertexBlendBoundaryGeometry::Circle {
                 curve: CurveGeometry::Nurbs(curve),
                 curve_endpoints,
@@ -1132,7 +1132,7 @@ fn decode_revision_vertex_blend_boundary(
                 tables,
             )?;
             let pcurve = decode_nullable_embedded_pcurve(bytes, position, int_width)?;
-            let sense = i64::from(take_bool(bytes, position)?);
+            let sense = take_bool(bytes, position)?;
             let fit_tolerance = take_f64(bytes, position)?;
             EmbeddedVertexBlendBoundaryGeometry::Pcurve {
                 surface: surface?,
